Move Workers to data layer
